requirements.txt
Flask==2.2.2
pymongo==4.2.0
dnspython==2.2.1
gunicorn==20.1.0
python-dotenv==0.20.0
Flask-WTF==1.0.1
email-validator==1.3.0
passlib==1.7.4